fix warning: Since PHP 7.2.0 count() yields a warning on invalid countable types such as null

+++ app/controllers/NewsController.php
@@ -125,7 +125,7 @@ class NewsController extends BaseController
 
         /* Delete the selected pictures */
         $picstodelete = Input::get('picstodelete');
-        if (count($picstodelete) > 0) {
+        if ($picstodelete !== null && count($picstodelete) > 0) {
             foreach($picstodelete as $picId) {
                 $pic = NewsPic::findOrFail($picId);
                 unlink($images_dir . $pic->filename); // "Delete" the file.
